requirements.txt
bs4
openai
chromadb
tiktoken
langchain
streamlit
streamlit-chat
streamlit-modal
psycopg2
python-dotenv
